chore: simplify Sentry and NextJS config
