--- sample-apps/react/react-dogfood/next.config.mjs
@@ -1,108 +1,0 @@
-// @ts-check
-
-import { withSentryConfig } from '@sentry/nextjs';
-import dotenv from 'dotenv';
-
-const env = dotenv.config({ path: '.env.local' });
-
-/**
- * @type {import('next').NextConfig}
- */
-const nextConfig = {
-  env: env.parsed,
-  productionBrowserSourceMaps: true,
-  basePath: process.env.NEXT_PUBLIC_BASE_PATH || '',
-  reactStrictMode: true,
-  reactCompiler: true,
-  async headers() {
-    return [
-      {
-        source: '/.well-known/apple-app-site-association',
-        headers: [
-          {
-            key: 'content-type',
-            value: 'application/json',
-          },
-        ],
-      },
-      {
-        source: '/api/auth/create-token',
-        headers: [
-          {
-            key: 'Access-Control-Allow-Origin',
-            value: '*',
-          },
-        ],
-      },
-      {
-        source: '/api/call/sample',
-        headers: [
-          {
-            key: 'Access-Control-Allow-Origin',
-            value: '*',
-          },
-        ],
-      },
-      {
-        source: '/(.*)',
-        headers: [
-          {
-            key: 'Content-Security-Policy',
-            value:
-              "frame-ancestors 'self' https://stream-generic-demo.vercel.app http://localhost:5100;",
-          },
-        ],
-      },
-    ];
-  },
-};
-
-/**
- * @type {import('@sentry/nextjs').SentryWebpackPluginOptions}
- */
-const sentryWebpackPluginOptions = {
-  // Additional config options for the Sentry Webpack plugin. Keep in mind that
-  // the following options are set automatically, and overriding them is not
-  // recommended:
-  //   release, url, org, project, authToken, configFile, stripPrefix,
-  //   urlPrefix, include, ignore
-
-  silent: true, // Suppresses all logs
-  // For all available options, see:
-  // https://github.com/getsentry/sentry-webpack-plugin#options.
-
-  // prevents creating Sentry releases on local builds.
-  dryRun: typeof process.env.CI === 'undefined',
-  authToken: process.env.PRONTO_SENTRY_AUTH_TOKEN,
-  org: 'stream',
-  project: 'video-pronto',
-  // For all available options, see:
-  // https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/
-
-  // Upload a larger set of source maps for prettier stack traces (increases build time)
-  widenClientFileUpload: true,
-
-  // Transpiles SDK to be compatible with IE11 (increases bundle size)
-  transpileClientSDK: false,
-
-  // Routes browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers. (increases server load)
-  // Note: Check that the configured route will not match with your Next.js middleware, otherwise reporting of client-
-  // side errors will fail.
-  tunnelRoute: '/sentry',
-
-  // Hides source maps from generated client bundles
-  hideSourceMaps: false,
-
-  // Automatically tree-shake Sentry logger statements to reduce bundle size
-  disableLogger: true,
-
-  // Enables automatic instrumentation of Vercel Cron Monitors.
-  // See the following for more information:
-  // https://docs.sentry.io/product/crons/
-  // https://vercel.com/docs/cron-jobs
-  automaticVercelMonitors: false,
-};
-
-// Make sure adding Sentry options is the last code to run before exporting, to
-// ensure that your source maps include changes from all other Webpack plugins
-export default withSentryConfig(nextConfig, sentryWebpackPluginOptions);
